refactor(pr-review): create src/pr-review/ boundary — circuit, authorization, critic-routing, persistence modules
- Move the pure v2 resilience circuit machine from src/hooks/pr-review-resilience-circuit.ts
  to src/pr-review/circuit.ts; add the sanctioned transitions that previously lived as inline
  record constructions in the gate (resolvePrReviewResiliencePolicy from config defaults,
  resetPrReviewResilienceForReEnable, rollbackPrReviewCircuitProbe) — recurrence classes
  G-1/G-2 from issue #2385.
- Consolidate the terminal-status vocabulary into CIRCUIT_TERMINAL_DELEGATION_STATUSES
  (single definition, G-6).
- Move reentry authorization to src/pr-review/authorization.ts with a bound
  binding-reader seam (no boundary->gate import).
- Move the critic-routing predicate canonically to src/pr-review/critic-routing.ts
  (contract re-exports; still exactly one definition).
- Extract the gate's persistence machinery (locks, CAS write, bounded parsing, atomic
  JSON writes, safe-parent/identity assertions, caches) into src/pr-review/persistence.ts
  with call-time-readable bound hooks and a bound state codec; verified byte-equivalent
  against the prior gate implementations.
